Made changes to the nav menu because the vertical split borders weren't reaching the top and bottom of the nav strip.  I had to adjust the padding top and bottom to stretch the element and the border to match.
In addition, any changes you want to make to the nav menu must also be copied or reflected to the mixins for midnight js.  I didn't realize this and wondered why my changes to the nav menu jade didn't do anything --- it was because I'd added data-type to the title.block jade and so Midnight JS took priority which meant I had to edit the mixin and not the original code.